mcp-server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests",
    "googlemaps>=4.10.0",
    "python-dotenv>=1.1.1",
    "langchain-groq>=0.3.7",
    "langchain-mcp-adapters>=0.1.9",
    "mcp>=1.13.0",
    "langgraph>=0.6.5",
]
